API/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
pymysql==1.0.2
python-dotenv==1.0.1
flask-cors==5.0.1
requests==2.32.3
cryptography==44.0.2